restrict access to manager/admin routes
