+++ client/src/components/ProtectedRoute.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import { INDEX_PATH, DASHBOARD_PATH } from '../constants/paths';
-import { useSelector, useDispatch } from 'react-redux';
-import { setAlert, AlertType } from '../actions/alerts';
+import { INDEX_PATH } from '../constants/paths';
+import { useSelector } from 'react-redux';
 
 export default function ProtectedRoute({ component: Component, ...rest }) {
   const auth = useSelector(state => state.auth);
@@ -17,30 +16,6 @@ export default function ProtectedRoute({ component: Component, ...rest }) {
           <Component {...props} />
         )
       }
-    />
-  );
-}
-
-export function OrgRoute({ component: Component, ...rest }) {
-  const dispatch = useDispatch();
-  const profile = useSelector(state => state.profile);
-
-  return (
-    <Route
-      {...rest}
-      render={props => {
-        if (profile.data.isAdmin || profile.data.isManager) {
-          return <Component {...props} />;
-        }
-
-        dispatch(
-          setAlert(
-            'You do not have access to the route you requested',
-            AlertType.WARNING
-          )
-        );
-        return <Redirect to={DASHBOARD_PATH} />;
-      }}
     />
   );
 }
