+++ client/src/pages/Dashboard.js
@@ -5,7 +5,6 @@ import {
   dashboardMenuItemPath,
   dashboardSchedulePath
 } from '../constants/paths';
-// FIXME: import { OrgRoute } from '../components/ProtectedRoute';
 
 import DashboardSidebar from '../components/dashboard/DashboardSidebar';
 import DashboardMenuItem from '../components/dashboard/DashboardMenuItem';
